pyproject.toml
[project]
name = "content-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "aiosmtplib>=4.0.1",
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.5.5",
    "fpdf>=1.7.2",
    "markdownify>=1.1.0",
    "openai-agents>=0.0.18",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic[email]>=2.11.7",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "requests==2.32.3",
    "rich>=14.0.0",
    "tavily-python>=0.7.6",
    "tenacity>=9.1.2",
]
